% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expldata.R
\name{expldata}
\alias{expldata}
\title{Explore a dataset}
\usage{
expldata(x)
}
\arguments{
\item{x}{A data frame with the folloging columns: \code{logL} and \code{invTemp}, which contain the log-likelihood and inverse temperature values, respectively.}
}
\value{
It produces a data.frame with the number of samples per temperature.
}
\description{
This function counts the number of samples per temperature.
}
\examples{
\dontrun{
data(ligoVirgoSim)
expldata(ligoVirgoSim)
}
}
\author{
Patricio Maturana Russel \email{p.russel@auckland.ac.nz}
}
