% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin}
\alias{burnin}
\title{Burnin period}
\usage{
burnin(x, percentage = NULL, nburnin = NULL)
}
\arguments{
\item{x}{A data frame with the folloging columns: \code{logL} and \code{invTemp}, which contain the log-likelihood and inverse temperature values, respectively.}

\item{percentage}{Percentage of log-likelihod values to discard at each temperature.}

\item{nburnin}{Number of log-likelihod values to discard at each temperature.}
}
\value{
It produces a new \code{x} with a burnin period.
}
\description{
This function discards the first log-likelihood values at each inverse temperature as burnin period.
}
\examples{
\dontrun{
data(ligoVirgoSim)
burnin(ligoVirgoSim, percentage=10); #discarding 10\\\% of log-likelihood values at each temperature
}
}
\author{
Patricio Maturana Russel \email{p.russel@auckland.ac.nz}
}
